+++ holypipette/__init__.py
@@ -1,0 +1,6 @@
+# Package initialization for holypipette
+# Re-export common utilities
+from .utils.config import Config, NumberWithUnit, Number, Boolean, Selector
+from .utils.log_utils import LoggingObject, setup_logging
+from .utils.exception_handler import set_global_exception_hook
+
