+++ holypipette/utils/__init__.py
@@ -1,3 +1,7 @@
 from .FileLogger import FileLogger
 from .EPhysLogger import EPhysLogger
 from .RecordingStateManager import RecordingStateManager
+from .config import Config, NumberWithUnit, Number, Boolean, Selector
+from .log_utils import LoggingObject, setup_logging
+from .exception_handler import set_global_exception_hook
+
